ocamlPackages.mdx: remove incorrect optimization
The original purpose of this optimization was described by its author
as follows:

| Strip optional dependencies from logs to make the closure smaller as this
| package contains a binary

However, that optimization is _incorrect_ because it affects not only
the command line tool (output "mdx.bin") but also the mdx library
(output "mdx").

Pull request #290291 recognized this issue, but merely moved this
optimization from development/ocaml-modules/mdx/default.nix to
top-level/ocaml-packages.nix, so everyone affected can help themselves
via attribute overriding.  As such, it didn't fix the original
problem, i.e. that this optimization is simply incorrect.  Moreover,
that doesn't help if we ever want an affected library, such as
"tls-eio", to be officially included in nixpkgs.

This commit fixes the root of the problem by removing the incorrect
optimization.

Please note that this commit is not intended to discourage the
re-introduction of such an optimization.  However, a new variant
of this optimization should only be included if it is correct,
i.e. reduces just "mdx.bin" without making "mdx" incompatible with
every package that directly or indirectly depends on the "logs"
library.